Add "representative" movie statistics.
- This prints a movie that represents each score/10 from 1 up to 10.
- This is calculated by assuming a gaussian distribution of scores, where the midpoint is between 5 and 6.
We expect then that a rating of 6 is any movie within 0.5 stdev above the midpoint, and 5 is 0.5 stdev below. 7 and 4
are 1 stdev either way, 8 and 3 are 1.5, 9 and 2 are 2stdev, and 10 and 1 are >2stddev.
The midpoint ranked movie of each range is found and returned as the representative of that score.